+++ lib/fastlane/plugin/wpmreleasetoolkit/actions/ios/ios_lint_localizations.rb
@@ -38,7 +38,7 @@ module Fastlane
           input_dir: resolve_path(params[:input_dir]),
           base_lang: params[:base_lang],
           only_langs: params[:only_langs],
-          fail_on_extra_strings: params[:fail_on_extra_strings]
+          fail_on_strings_not_in_base_language: params[:fail_on_strings_not_in_base_language]
         )
       end
 
@@ -185,8 +185,8 @@ module Fastlane
             type: Boolean
           ),
           FastlaneCore::ConfigItem.new(
-            key: :fail_on_extra_strings,
-            env_name: 'FL_IOS_LINT_TRANSLATIONS_FAIL_ON_EXTRA_STRINGS',
+            key: :fail_on_strings_not_in_base_language,
+            env_name: 'FL_IOS_LINT_TRANSLATIONS_FAIL_ON_STRINGS_NOT_IN_BASE_LANGUAGE',
             description: 'Should we report violations when finding strings in translations that are not present in the base language',
             optional: true,
             default_value: true,